+++ helper.py
@@ -1,3 +1,7 @@
+import json
+import subprocess
+
+
 def read_file(path):
     assert isinstance(path, str), "Path must be a string"
 
@@ -9,3 +13,20 @@ def read_file(path):
         return "Error: File not found."
     except Exception as e:
         return f"Error: {str(e)}"
+
+
+def get_video_duration_from_file(video_path):
+    try:
+        result = subprocess.run(
+            ['ffprobe', '-v', 'error', '-show_format', '-show_streams', '-of', 'json', video_path],
+            stdout=subprocess.PIPE, stderr=subprocess.PIPE
+        )
+
+        video_info = json.loads(result.stdout)
+
+        duration = float(video_info['format']['duration'])
+
+        return duration
+    except Exception as e:
+        print(f"Can't fetch Video-Duration: {str(e)}")
+        return 0
