+++ requirements.txt
@@ -1,0 +1,3 @@
+Flask==3.1.0
+Requests==2.32.3
+yt_dlp==2024.12.23
